Added support for opponents of a team
For most scenarios we want to generate T v T data for all the listed teams vs all the other listed teams. However there are times where we only want to generate a subset of the data. For example OU vs CU and then OU vs all the counties and CU vs all the counties but not all the counties vs each other, this change supports that